requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
Pillow
pytesseract
torch
torchvision
torchaudio
transformers
sentencepiece
numpy